refactor: peer manager, sink, stream, messages and fused
