+++ packages/peer/src/lib.rs
@@ -9,10 +9,11 @@ mod protocol;
 pub use codec::PeerProtocolCodec;
 
 pub use crate::manager::builder::PeerManagerBuilder;
+pub use crate::manager::messages::{IPeerManagerMessage, ManagedMessage, MessageId, OPeerManagerMessage};
 pub use crate::manager::peer_info::PeerInfo;
-pub use crate::manager::{
-    IPeerManagerMessage, ManagedMessage, MessageId, OPeerManagerMessage, PeerManager, PeerManagerSink, PeerManagerStream,
-};
+pub use crate::manager::sink::PeerManagerSink;
+pub use crate::manager::stream::PeerManagerStream;
+pub use crate::manager::PeerManager;
 pub use crate::protocol::{NestedPeerProtocol, PeerProtocol};
 
 /// Serializable and deserializable protocol messages.
