+++ packages/peer/src/manager/fused.rs
